src/server/requirements.txt
Flask
pathlib
flask_sqlalchemy
py-redis
